requirements.txt
pandas
transformers
streamlit
matplotlib
torch
sentencepiece
plotly
scipy
nltk
spacy
sumy
streamlit-toggle-switch